+++ src/App.js
@@ -240,7 +240,7 @@ class App extends Component {
           
           { !this.state.moreInfo ? <button onClick={this.showMoreInfo}>More info...</button> :
             <div>
-              <p>Enemy spawn locations grow in difficulty as you proceed deeper into the maze, as do gift box rewards (if randomize gift box contents is on, which is recommended). You will not have to move any NPC or do any storyline events, like riding the Sky Runner or beating Carpainter, to beat the game in this mode. Often, you can move an NPC or do story events to "skip" around in the cave. This may take you onto a different floor into the maze. If you take no skips, you will encounter each of the 8 shiny spots before reaching Giygas. The Exit Mouse will return you to the last sanctuary you visited.</p>
+              <p>Enemy spawn locations grow in difficulty as you proceed deeper into the maze, as do gift box rewards (if randomize gift box contents is on, which is recommended). You will not have to move any NPC or do any storyline events, like riding the Sky Runner or beating Carpainter, to beat the game in this mode. Often, you can move an NPC or do story events to "skip" around in the cave. This may take you onto a different floor into the maze. If you take no skips, you will encounter each of the 8 shiny spots before reaching Giygas. The Exit Mouse can remember a spot, and return you to that spot later.</p>
           
               <p>There is a <a href="https://github.com/pickfifteen/eb-randomizer/blob/master/ancient_cave_tips.md">document with strategies and tips</a> available.</p>