+++ src/Randomizer/Script.js
@@ -127,7 +127,6 @@ class Script {
         console.assert(this.lines[1][0] === 0x1b);
         this.lines = this.lines.slice(2);
         this.lines.unshift([0x05, 0x0B, 0x00]);   // encounters on
-        this.lines.unshift([0x1F, 0x68]);         // exit mouse
         console.assert(this.lines[this.lines.length-1][0] === 0x02);
         const keys = [[0x1F, 0x41, 0x05],];       // remove OSS-on
         this.removeInstructions(keys, []);